Prevent multiple updates after email verified already
