+++ src/backend/TrafficCourts/Workflow.Service/Sagas/VerifyEmailAddressSagaStateMachine.cs
@@ -253,52 +253,4 @@ public partial class VerifyEmailAddressStateMachine : MassTransitStateMachine<Ve
 
     [LoggerMessage(Level = LogLevel.Warning, Message = "Cannot change dispute id")]
     private partial void LogInvalidDisputeId([TagProvider(typeof(TagProvider), nameof(TagProvider.RecordTags), OmitReferenceName = true)] BehaviorContext<VerifyEmailAddressState, RequestEmailVerification> context);
-
-
-    internal static class TagProvider
-    {
-        public static void RecordTags(ITagCollector collector, BehaviorContext<VerifyEmailAddressState, SubmitNoticeOfDispute> context)
-        {
-            collector.Add(nameof(context.Message.NoticeOfDisputeGuid), context.Message.NoticeOfDisputeGuid);
-        }
-
-        public static void RecordTags(ITagCollector collector, BehaviorContext<VerifyEmailAddressState, RequestEmailVerification> context)
-        {
-            var state = context.Saga;
-            var message = context.Message;
-
-            collector.Add(nameof(message.NoticeOfDisputeGuid), message.NoticeOfDisputeGuid);
-        }
-
-        public static void RecordTags(ITagCollector collector, BehaviorContext<VerifyEmailAddressState, CheckEmailVerificationTokenRequest> context)
-        {
-            collector.Add(nameof(context.Message.NoticeOfDisputeGuid), context.Message.NoticeOfDisputeGuid);
-        }
-
-        public static void RecordTags(ITagCollector collector, ConsumeContext<CheckEmailVerificationTokenRequest> context)
-        {
-            collector.Add(nameof(context.Message.NoticeOfDisputeGuid), context.Message.NoticeOfDisputeGuid);
-        }
-
-        public static void RecordTags(ITagCollector collector, BehaviorContext<VerifyEmailAddressState, EmailVerificationSuccessful> context)
-        {
-            collector.Add(nameof(context.Message.NoticeOfDisputeGuid), context.Message.NoticeOfDisputeGuid);
-        }
-
-        public static void RecordTags(ITagCollector collector, BehaviorContext<VerifyEmailAddressState, ResendEmailVerificationEmail> context)
-        {
-            collector.Add(nameof(context.Message.NoticeOfDisputeGuid), context.Message.NoticeOfDisputeGuid);
-        }
-
-        public static void RecordTags(ITagCollector collector, BehaviorContext<VerifyEmailAddressState, NoticeOfDisputeSubmitted> context)
-        {
-            collector.Add(nameof(context.Message.NoticeOfDisputeGuid), context.Message.NoticeOfDisputeGuid);
-        }
-
-        public static void RecordTags(ITagCollector collector, BehaviorContext<VerifyEmailAddressState, SendEmailVerificationFailed> context)
-        {
-            collector.Add(nameof(context.Message.NoticeOfDisputeGuid), context.Message.NoticeOfDisputeGuid);
-            collector.Add(nameof(context.Message.Reason), context.Message.Reason);
-        }
-    }
 }
